+++ src/core/Bridge.ts
