+++ src/core/atom.ts
