+++ src/index.ts
@@ -1,5 +1,5 @@
-export { Provider } from './Provider'
-export { atom } from './atom'
-export { useAtom } from './useAtom'
-export type { Atom, WritableAtom, PrimitiveAtom } from './types'
-export { useBridge, Bridge } from './Bridge'
+export { Provider } from './core/Provider'
+export { atom } from './core/atom'
+export { useAtom } from './core/useAtom'
+export type { Atom, WritableAtom, PrimitiveAtom } from './core/types'
+export { useBridge, Bridge } from './core/Bridge'
