+++ src/utils/useUpdateAtom.ts
@@ -1,0 +1,12 @@
+import { useMemo } from 'react'
+import { atom, useAtom, WritableAtom } from 'jotai'
+
+export const useUpdateAtom = <Value, Update>(
+  anAtom: WritableAtom<Value, Update>
+) => {
+  const writeOnlyAtom = useMemo(
+    () => atom(null, (_get, set, update: Update) => set(anAtom, update)),
+    [anAtom]
+  )
+  return useAtom(writeOnlyAtom)[1]
+}
